final cleaning

+++ app/src/main/java/xyz/sevive/arcaeaoffline/ui/activities/EmergencyModeActivityViewModel.kt
@@ -90,9 +90,12 @@ class EmergencyModeActivityViewModel(
 
     fun deleteAllOcrDependencies() {
         val paths = OcrDependencyPaths()
-        SystemFileSystem.delete(paths.knnModelFile)
-        SystemFileSystem.delete(paths.phashDatabaseFile)
-        SystemFileSystem.delete(paths.imageHashesDatabaseFile)
+
+        viewModelScope.launch(Dispatchers.IO) {
+            if (SystemFileSystem.exists(paths.knnModelFile)) SystemFileSystem.delete(paths.knnModelFile)
+            if (SystemFileSystem.exists(paths.phashDatabaseFile)) SystemFileSystem.delete(paths.phashDatabaseFile)
+            if (SystemFileSystem.exists(paths.imageHashesDatabaseFile)) SystemFileSystem.delete(paths.imageHashesDatabaseFile)
+        }
     }
 
     fun deleteOcrQueueDatabase(context: Context) {
